Support multiple sensu servers/backends (#36)
Modifies the k8s StatefulSet of sensu backends to support multiple
pods. Changed the sensu config initContainer build etcd flags based on
the ordinal number of the pod. Changed the etcd data PVC to be
allocated in a volume claim template, so PVCs are named with ordinal
numbers by the StatefulSet.

Scaling is handled by incrementing or decrementing the number of
replicas in the StatefulSet one at a time, and adding/removing a
member from etcd until the desired size is reached. Upgrades are
handled by updating the StatefulSet with the new sensu image version
and letting the StatefulSet rolling update occur.

Also updated k8s calls from using AppsV1Beta1 to AppsV1 to be
compatible with k8s 1.16.